+++ ThirdPartyAdapters/mopub/CHANGELOG.md
@@ -1,5 +1,13 @@
 # MoPub Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 4.15.0.0
+- Verified compatibility with MoPub SDK 4.15.0.
+
+## 4.14.0.0
+- Verified compatibility with MoPub SDK 4.14.0.
+- Adapter shows a warning when the requested banner ad size does not match the
+  ad size set in MoPub UI.
+
 ## 4.13.0.0
 - Verified compatibility with MoPub SDK 4.13.0.